DNSSEC: Handling of SOA RRSIGs.
- Quite a lot of changes had to be done, because some variables were only accesible locally.
- Some includes might not be needed, needs a second look.

Refs #4

+++ src/zscanner/scanner.c
@@ -5185,7 +5185,7 @@ _resume:
 			else if ( _widec > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (unsigned int)(_mid - _keys);
+				_trans += (_mid - _keys);
 				goto _match;
 			}
 		}
@@ -5208,7 +5208,7 @@ _resume:
 			else if ( _widec > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += (unsigned int)((_mid - _keys)>>1);
+				_trans += ((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
